+++ Validation-Manager-Web/src/main/java/com/validation/manager/core/db/controller/RequirementTypeJpaController.java
@@ -4,14 +4,17 @@
  */
 package com.validation.manager.core.db.controller;
 
+import com.validation.manager.core.db.Requirement;
 import com.validation.manager.core.db.RequirementType;
+import com.validation.manager.core.db.controller.exceptions.IllegalOrphanException;
 import com.validation.manager.core.db.controller.exceptions.NonexistentEntityException;
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
+import javax.persistence.Query;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 
@@ -31,11 +34,29 @@ public class RequirementTypeJpaController implements Serializable {
     }
 
     public void create(RequirementType requirementType) {
+        if (requirementType.getRequirementList() == null) {
+            requirementType.setRequirementList(new ArrayList<Requirement>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Requirement> attachedRequirementList = new ArrayList<Requirement>();
+            for (Requirement requirementListRequirementToAttach : requirementType.getRequirementList()) {
+                requirementListRequirementToAttach = em.getReference(requirementListRequirementToAttach.getClass(), requirementListRequirementToAttach.getRequirementPK());
+                attachedRequirementList.add(requirementListRequirementToAttach);
+            }
+            requirementType.setRequirementList(attachedRequirementList);
             em.persist(requirementType);
+            for (Requirement requirementListRequirement : requirementType.getRequirementList()) {
+                RequirementType oldRequirementTypeOfRequirementListRequirement = requirementListRequirement.getRequirementType();
+                requirementListRequirement.setRequirementType(requirementType);
+                requirementListRequirement = em.merge(requirementListRequirement);
+                if (oldRequirementTypeOfRequirementListRequirement != null) {
+                    oldRequirementTypeOfRequirementListRequirement.getRequirementList().remove(requirementListRequirement);
+                    oldRequirementTypeOfRequirementListRequirement = em.merge(oldRequirementTypeOfRequirementListRequirement);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -44,12 +65,45 @@ public class RequirementTypeJpaController implements Serializable {
         }
     }
 
-    public void edit(RequirementType requirementType) throws NonexistentEntityException, Exception {
+    public void edit(RequirementType requirementType) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            RequirementType persistentRequirementType = em.find(RequirementType.class, requirementType.getId());
+            List<Requirement> requirementListOld = persistentRequirementType.getRequirementList();
+            List<Requirement> requirementListNew = requirementType.getRequirementList();
+            List<String> illegalOrphanMessages = null;
+            for (Requirement requirementListOldRequirement : requirementListOld) {
+                if (!requirementListNew.contains(requirementListOldRequirement)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Requirement " + requirementListOldRequirement + " since its requirementType field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Requirement> attachedRequirementListNew = new ArrayList<Requirement>();
+            for (Requirement requirementListNewRequirementToAttach : requirementListNew) {
+                requirementListNewRequirementToAttach = em.getReference(requirementListNewRequirementToAttach.getClass(), requirementListNewRequirementToAttach.getRequirementPK());
+                attachedRequirementListNew.add(requirementListNewRequirementToAttach);
+            }
+            requirementListNew = attachedRequirementListNew;
+            requirementType.setRequirementList(requirementListNew);
             requirementType = em.merge(requirementType);
+            for (Requirement requirementListNewRequirement : requirementListNew) {
+                if (!requirementListOld.contains(requirementListNewRequirement)) {
+                    RequirementType oldRequirementTypeOfRequirementListNewRequirement = requirementListNewRequirement.getRequirementType();
+                    requirementListNewRequirement.setRequirementType(requirementType);
+                    requirementListNewRequirement = em.merge(requirementListNewRequirement);
+                    if (oldRequirementTypeOfRequirementListNewRequirement != null && !oldRequirementTypeOfRequirementListNewRequirement.equals(requirementType)) {
+                        oldRequirementTypeOfRequirementListNewRequirement.getRequirementList().remove(requirementListNewRequirement);
+                        oldRequirementTypeOfRequirementListNewRequirement = em.merge(oldRequirementTypeOfRequirementListNewRequirement);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -67,7 +121,7 @@ public class RequirementTypeJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -78,6 +132,17 @@ public class RequirementTypeJpaController implements Serializable {
                 requirementType.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The requirementType with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Requirement> requirementListOrphanCheck = requirementType.getRequirementList();
+            for (Requirement requirementListOrphanCheckRequirement : requirementListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This RequirementType (" + requirementType + ") cannot be destroyed since the Requirement " + requirementListOrphanCheckRequirement + " in its requirementList field has a non-nullable requirementType field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(requirementType);
             em.getTransaction().commit();
